+++ simplegen-api/src/main/kotlin/com/ancientlightstudios/simplegen/DataParser.kt
@@ -15,6 +15,11 @@ interface DataParser {
     val supportedDataFormats: Set<String>
 
     /**
+     * The default path in the resulting data map where the parsed data should be stored.
+     */
+    val defaultResultPath: String get() = ""
+
+    /**
      * Initialize the parser with the given configuration. The configuration is parser specific.
      */
     fun init(configuration: Map<String,Any>)
@@ -31,4 +36,47 @@ interface DataParser {
      * @exception DataParseException in case the data cannot be parsed.
      */
     fun parse(stream: InputStream, origin: String, configuration: Map<String, Any>): Map<String, Any>
+
+
+    /**
+     * Extracts the first character of the given value. If the value is no string, null or empty, the
+     * default value is returned.
+     */
+    fun firstCharOfStringOrDefault(value: Any?, defaultValue: Char): Char {
+        if (value is String) {
+            return value.firstOrNull() ?: defaultValue
+        }
+        if (value is Char) {
+            return value
+        }
+        return defaultValue
+    }
+
+    /**
+     * Converts the given value to an integer. If the value is no number, the default value is returned.
+     */
+    fun asIntOrDefault(value: Any?, defaultValue: Int): Int {
+        if (value is Int) {
+            return value
+        }
+        if (value is String) {
+            return value.toIntOrNull() ?: defaultValue
+        }
+        return defaultValue
+    }
+
+    /**
+     * Converts the given value to a boolean. If the value is no boolean, it will be checked if it is a string.
+     * If it is a string, it will be checked if it is "true" or "false". If it is neither a boolean nor a string,
+     * the default value is returned.
+     */
+    fun booleanOrDefault(value: Any?, defaultValue: Boolean): Boolean {
+        if (value is Boolean) {
+            return value
+        }
+        if (value is String) {
+            return value.toBoolean()
+        }
+        return defaultValue
+    }
 }