+++ simplegen-dataformat-csv/src/test/kotlin/com/ancientlightstudios/simplegen/CsvParserTests.kt
@@ -23,10 +23,10 @@ class CsvParserTests : io.kotest.core.spec.style.BehaviorSpec({
             val result = underTest.parse(data.byteInputStream(), "plain text", mapOf())
 
             Then("the data is properly parsed") {
-                result shouldHaveKey "csv"
+                result shouldHaveKey "entries"
 
                 @Suppress("UNCHECKED_CAST")
-                val inner = result["csv"] as List<Map<String, Any>>
+                val inner = result["entries"] as List<Map<String, Any>>
                 inner.size shouldBe 2
                 inner[0]["name"] shouldBe "john"
                 inner[0]["age"] shouldBe "42"
@@ -49,27 +49,20 @@ class CsvParserTests : io.kotest.core.spec.style.BehaviorSpec({
 
             val settings = mapOf(
                 "skipLines" to 4,
-                "resultPath" to "some.nested.path",
                 "quoteChar" to "'"
             )
             val result = underTest.parse(data.byteInputStream(), "plain text", settings)
 
             Then("the data is properly parsed") {
-                result shouldHaveKey "some"
+                result shouldHaveKey "entries"
 
                 @Suppress("UNCHECKED_CAST")
-                val inner = result["some"] as Map<String, Any>
-                inner shouldHaveKey "nested"
-                @Suppress("UNCHECKED_CAST")
-                val inner2 = inner["nested"] as Map<String, Any>
-                inner2 shouldHaveKey "path"
-                @Suppress("UNCHECKED_CAST")
-                val inner3 = inner2["path"] as List<Map<String, Any>>
-                inner3.size shouldBe 2
-                inner3[0]["name"] shouldBe "john"
-                inner3[0]["age"] shouldBe "42"
-                inner3[1]["name"] shouldBe "jane"
-                inner3[1]["age"] shouldBe "43"
+                val entries = result["entries"] as List<Map<String, Any>>
+                entries.size shouldBe 2
+                entries[0]["name"] shouldBe "john"
+                entries[0]["age"] shouldBe "42"
+                entries[1]["name"] shouldBe "jane"
+                entries[1]["age"] shouldBe "43"
             }
         }