+++ simplegen/src/main/kotlin/com/ancientlightstudios/simplegen/DataLoader.kt
@@ -3,7 +3,6 @@ package com.ancientlightstudios.simplegen
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.File
-import java.io.InputStream
 import java.util.*
 
 /**
@@ -38,21 +37,38 @@ object DataLoader {
         parsersByMimeType = tempMap
     }
 
-    fun parse(file: File, parserSettings: Map<String, Any>, mimeType: String? = null): Map<String, Any> {
+    fun parse(file: File, mimeType: String? = null, resultPath: String, parserSettings: Map<String, Any>): Map<String, Any> {
         val path = file.path
         val realMimeType = mimeType
             ?: "application/yaml"
 
         return file.inputStream().use {
-            parse(it, path, realMimeType, parserSettings)
+            val parser = parsersByMimeType[realMimeType]
+                ?: throw DataParseException(path, "No parser is available for mime type $realMimeType")
+            val parseResult = parser.parse(it, path, parserSettings)
+            if (resultPath.isEmpty()) {
+                return@use parseResult
+            }
+            // the result will be a list of maps, each map representing a row
+            // the user can configure the path of the result in the output map
+
+            // we split the path into its parts and create nested maps for each part
+            val resultPathParts = resultPath.split(".")
+            val result = mutableMapOf<String, Any>()
+
+            // walk the path parts and create the nested maps
+            var currentMap = result
+            for (i in 0 until resultPathParts.size - 1) { // we don't need the last part, as it will be the key of the last map
+                val part = resultPathParts[i]
+                val newMap = mutableMapOf<String, Any>()
+                currentMap[part] = newMap
+                currentMap = newMap
+            }
+
+            // and add it to the last map
+            currentMap[resultPathParts.last()] = parseResult
+            result
         }
-    }
-
-    fun parse(inputStream: InputStream, origin: String, mimeType: String, parserSettings: Map<String, Any>): Map<String, Any> {
-        val parser = parsersByMimeType[mimeType]
-            ?: throw DataParseException(origin, "No parser is available for mime type $mimeType")
-
-        return parser.parse(inputStream, origin, parserSettings)
     }
 
     /**